Rest

+++ Memberships/Areas/Admin/Controllers/ItemTypeController.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.Data;
 using System.Data.Entity;
 using System.Linq;
+using System.Threading.Tasks;
 using System.Net;
 using System.Web;
 using System.Web.Mvc;
@@ -11,24 +12,25 @@ using Memberships.Models;
 
 namespace Memberships.Areas.Admin.Controllers
 {
+    [Authorize(Roles = "Admin")]
     public class ItemTypeController : Controller
     {
         private ApplicationDbContext db = new ApplicationDbContext();
 
         // GET: Admin/ItemType
-        public ActionResult Index()
+        public async Task<ActionResult> Index()
         {
-            return View(db.ItemTypes.ToList());
+            return View(await db.ItemTypes.ToListAsync());
         }
 
         // GET: Admin/ItemType/Details/5
-        public ActionResult Details(int? id)
+        public async Task<ActionResult> Details(int? id)
         {
             if (id == null)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
             }
-            ItemType itemType = db.ItemTypes.Find(id);
+            ItemType itemType = await db.ItemTypes.FindAsync(id);
             if (itemType == null)
             {
                 return HttpNotFound();
@@ -44,15 +46,15 @@ namespace Memberships.Areas.Admin.Controllers
 
         // POST: Admin/ItemType/Create
         // To protect from overposting attacks, please enable the specific properties you want to bind to, for 
-        // more details see https://go.microsoft.com/fwlink/?LinkId=317598.
+        // more details see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public ActionResult Create([Bind(Include = "Id,Title")] ItemType itemType)
+        public async Task<ActionResult> Create([Bind(Include = "Id,Title")] ItemType itemType)
         {
             if (ModelState.IsValid)
             {
                 db.ItemTypes.Add(itemType);
-                db.SaveChanges();
+                await db.SaveChangesAsync();
                 return RedirectToAction("Index");
             }
 
@@ -60,13 +62,13 @@ namespace Memberships.Areas.Admin.Controllers
         }
 
         // GET: Admin/ItemType/Edit/5
-        public ActionResult Edit(int? id)
+        public async Task<ActionResult> Edit(int? id)
         {
             if (id == null)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
             }
-            ItemType itemType = db.ItemTypes.Find(id);
+            ItemType itemType = await db.ItemTypes.FindAsync(id);
             if (itemType == null)
             {
                 return HttpNotFound();
@@ -76,28 +78,28 @@ namespace Memberships.Areas.Admin.Controllers
 
         // POST: Admin/ItemType/Edit/5
         // To protect from overposting attacks, please enable the specific properties you want to bind to, for 
-        // more details see https://go.microsoft.com/fwlink/?LinkId=317598.
+        // more details see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public ActionResult Edit([Bind(Include = "Id,Title")] ItemType itemType)
+        public async Task<ActionResult> Edit([Bind(Include = "Id,Title")] ItemType itemType)
         {
             if (ModelState.IsValid)
             {
                 db.Entry(itemType).State = EntityState.Modified;
-                db.SaveChanges();
+                await db.SaveChangesAsync();
                 return RedirectToAction("Index");
             }
             return View(itemType);
         }
 
         // GET: Admin/ItemType/Delete/5
-        public ActionResult Delete(int? id)
+        public async Task<ActionResult> Delete(int? id)
         {
             if (id == null)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
             }
-            ItemType itemType = db.ItemTypes.Find(id);
+            ItemType itemType = await db.ItemTypes.FindAsync(id);
             if (itemType == null)
             {
                 return HttpNotFound();
@@ -108,11 +110,16 @@ namespace Memberships.Areas.Admin.Controllers
         // POST: Admin/ItemType/Delete/5
         [HttpPost, ActionName("Delete")]
         [ValidateAntiForgeryToken]
-        public ActionResult DeleteConfirmed(int id)
+        public async Task<ActionResult> DeleteConfirmed(int id)
         {
-            ItemType itemType = db.ItemTypes.Find(id);
-            db.ItemTypes.Remove(itemType);
-            db.SaveChanges();
+            ItemType itemType = await db.ItemTypes.FindAsync(id);
+            var isUnused = await db.Items.CountAsync(i => i.ItemTypeId.Equals(id)) == 0;
+            if (isUnused)
+            {
+                db.ItemTypes.Remove(itemType);
+                await db.SaveChangesAsync();
+            }
+
             return RedirectToAction("Index");
         }
