+++ Memberships/Areas/Admin/Controllers/PartController.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.Data;
 using System.Data.Entity;
 using System.Linq;
+using System.Threading.Tasks;
 using System.Net;
 using System.Web;
 using System.Web.Mvc;
@@ -11,24 +12,25 @@ using Memberships.Models;
 
 namespace Memberships.Areas.Admin.Controllers
 {
+    [Authorize(Roles = "Admin")]
     public class PartController : Controller
     {
         private ApplicationDbContext db = new ApplicationDbContext();
 
         // GET: Admin/Part
-        public ActionResult Index()
+        public async Task<ActionResult> Index()
         {
-            return View(db.Parts.ToList());
+            return View(await db.Parts.ToListAsync());
         }
 
         // GET: Admin/Part/Details/5
-        public ActionResult Details(int? id)
+        public async Task<ActionResult> Details(int? id)
         {
             if (id == null)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
             }
-            Part part = db.Parts.Find(id);
+            Part part = await db.Parts.FindAsync(id);
             if (part == null)
             {
                 return HttpNotFound();
@@ -44,15 +46,15 @@ namespace Memberships.Areas.Admin.Controllers
 
         // POST: Admin/Part/Create
         // To protect from overposting attacks, please enable the specific properties you want to bind to, for 
-        // more details see https://go.microsoft.com/fwlink/?LinkId=317598.
+        // more details see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public ActionResult Create([Bind(Include = "Id,Title")] Part part)
+        public async Task<ActionResult> Create([Bind(Include = "Id,Title")] Part part)
         {
             if (ModelState.IsValid)
             {
                 db.Parts.Add(part);
-                db.SaveChanges();
+                await db.SaveChangesAsync();
                 return RedirectToAction("Index");
             }
 
@@ -60,13 +62,13 @@ namespace Memberships.Areas.Admin.Controllers
         }
 
         // GET: Admin/Part/Edit/5
-        public ActionResult Edit(int? id)
+        public async Task<ActionResult> Edit(int? id)
         {
             if (id == null)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
             }
-            Part part = db.Parts.Find(id);
+            Part part = await db.Parts.FindAsync(id);
             if (part == null)
             {
                 return HttpNotFound();
@@ -76,28 +78,28 @@ namespace Memberships.Areas.Admin.Controllers
 
         // POST: Admin/Part/Edit/5
         // To protect from overposting attacks, please enable the specific properties you want to bind to, for 
-        // more details see https://go.microsoft.com/fwlink/?LinkId=317598.
+        // more details see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public ActionResult Edit([Bind(Include = "Id,Title")] Part part)
+        public async Task<ActionResult> Edit([Bind(Include = "Id,Title")] Part part)
         {
             if (ModelState.IsValid)
             {
                 db.Entry(part).State = EntityState.Modified;
-                db.SaveChanges();
+                await db.SaveChangesAsync();
                 return RedirectToAction("Index");
             }
             return View(part);
         }
 
         // GET: Admin/Part/Delete/5
-        public ActionResult Delete(int? id)
+        public async Task<ActionResult> Delete(int? id)
         {
             if (id == null)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
             }
-            Part part = db.Parts.Find(id);
+            Part part = await db.Parts.FindAsync(id);
             if (part == null)
             {
                 return HttpNotFound();
@@ -108,11 +110,16 @@ namespace Memberships.Areas.Admin.Controllers
         // POST: Admin/Part/Delete/5
         [HttpPost, ActionName("Delete")]
         [ValidateAntiForgeryToken]
-        public ActionResult DeleteConfirmed(int id)
+        public async Task<ActionResult> DeleteConfirmed(int id)
         {
-            Part part = db.Parts.Find(id);
-            db.Parts.Remove(part);
-            db.SaveChanges();
+            Part part = await db.Parts.FindAsync(id);
+            var isUnused = await db.Items.CountAsync(i => i.PartId.Equals(id)) == 0;
+            if (isUnused)
+            {
+                db.Parts.Remove(part);
+                await db.SaveChangesAsync();
+            }
+
             return RedirectToAction("Index");
         }
