+++ Memberships/Areas/Admin/Controllers/ProductTypeController.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.Data;
 using System.Data.Entity;
 using System.Linq;
+using System.Threading.Tasks;
 using System.Net;
 using System.Web;
 using System.Web.Mvc;
@@ -11,24 +12,25 @@ using Memberships.Models;
 
 namespace Memberships.Areas.Admin.Controllers
 {
+    [Authorize(Roles = "Admin")]
     public class ProductTypeController : Controller
     {
         private ApplicationDbContext db = new ApplicationDbContext();
 
         // GET: Admin/ProductType
-        public ActionResult Index()
+        public async Task<ActionResult> Index()
         {
-            return View(db.ProductTypes.ToList());
+            return View(await db.ProductTypes.ToListAsync());
         }
 
         // GET: Admin/ProductType/Details/5
-        public ActionResult Details(int? id)
+        public async Task<ActionResult> Details(int? id)
         {
             if (id == null)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
             }
-            ProductType productType = db.ProductTypes.Find(id);
+            ProductType productType = await db.ProductTypes.FindAsync(id);
             if (productType == null)
             {
                 return HttpNotFound();
@@ -44,15 +46,15 @@ namespace Memberships.Areas.Admin.Controllers
 
         // POST: Admin/ProductType/Create
         // To protect from overposting attacks, please enable the specific properties you want to bind to, for 
-        // more details see https://go.microsoft.com/fwlink/?LinkId=317598.
+        // more details see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public ActionResult Create([Bind(Include = "Id,Title")] ProductType productType)
+        public async Task<ActionResult> Create([Bind(Include = "Id,Title")] ProductType productType)
         {
             if (ModelState.IsValid)
             {
                 db.ProductTypes.Add(productType);
-                db.SaveChanges();
+                await db.SaveChangesAsync();
                 return RedirectToAction("Index");
             }
 
@@ -60,13 +62,13 @@ namespace Memberships.Areas.Admin.Controllers
         }
 
         // GET: Admin/ProductType/Edit/5
-        public ActionResult Edit(int? id)
+        public async Task<ActionResult> Edit(int? id)
         {
             if (id == null)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
             }
-            ProductType productType = db.ProductTypes.Find(id);
+            ProductType productType = await db.ProductTypes.FindAsync(id);
             if (productType == null)
             {
                 return HttpNotFound();
@@ -76,28 +78,28 @@ namespace Memberships.Areas.Admin.Controllers
 
         // POST: Admin/ProductType/Edit/5
         // To protect from overposting attacks, please enable the specific properties you want to bind to, for 
-        // more details see https://go.microsoft.com/fwlink/?LinkId=317598.
+        // more details see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public ActionResult Edit([Bind(Include = "Id,Title")] ProductType productType)
+        public async Task<ActionResult> Edit([Bind(Include = "Id,Title")] ProductType productType)
         {
             if (ModelState.IsValid)
             {
                 db.Entry(productType).State = EntityState.Modified;
-                db.SaveChanges();
+                await db.SaveChangesAsync();
                 return RedirectToAction("Index");
             }
             return View(productType);
         }
 
         // GET: Admin/ProductType/Delete/5
-        public ActionResult Delete(int? id)
+        public async Task<ActionResult> Delete(int? id)
         {
             if (id == null)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
             }
-            ProductType productType = db.ProductTypes.Find(id);
+            ProductType productType = await db.ProductTypes.FindAsync(id);
             if (productType == null)
             {
                 return HttpNotFound();
@@ -108,11 +110,16 @@ namespace Memberships.Areas.Admin.Controllers
         // POST: Admin/ProductType/Delete/5
         [HttpPost, ActionName("Delete")]
         [ValidateAntiForgeryToken]
-        public ActionResult DeleteConfirmed(int id)
+        public async Task<ActionResult> DeleteConfirmed(int id)
         {
-            ProductType productType = db.ProductTypes.Find(id);
-            db.ProductTypes.Remove(productType);
-            db.SaveChanges();
+            ProductType productType = await db.ProductTypes.FindAsync(id);
+            var isUnused = await db.Products.CountAsync(i => i.ProductTypeId.Equals(id)) == 0;
+            if (isUnused)
+            {
+                db.ProductTypes.Remove(productType);
+                await db.SaveChangesAsync();
+            }
+
             return RedirectToAction("Index");
         }
