+++ Memberships/Areas/Admin/Controllers/SectionController.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.Data;
 using System.Data.Entity;
 using System.Linq;
+using System.Threading.Tasks;
 using System.Net;
 using System.Web;
 using System.Web.Mvc;
@@ -11,24 +12,25 @@ using Memberships.Models;
 
 namespace Memberships.Areas.Admin.Controllers
 {
+    [Authorize(Roles = "Admin")]
     public class SectionController : Controller
     {
         private ApplicationDbContext db = new ApplicationDbContext();
 
         // GET: Admin/Section
-        public ActionResult Index()
+        public async Task<ActionResult> Index()
         {
-            return View(db.Sections.ToList());
+            return View(await db.Sections.ToListAsync());
         }
 
         // GET: Admin/Section/Details/5
-        public ActionResult Details(int? id)
+        public async Task<ActionResult> Details(int? id)
         {
             if (id == null)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
             }
-            Section section = db.Sections.Find(id);
+            Section section = await db.Sections.FindAsync(id);
             if (section == null)
             {
                 return HttpNotFound();
@@ -44,15 +46,16 @@ namespace Memberships.Areas.Admin.Controllers
 
         // POST: Admin/Section/Create
         // To protect from overposting attacks, please enable the specific properties you want to bind to, for 
-        // more details see https://go.microsoft.com/fwlink/?LinkId=317598.
+        // more details see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public ActionResult Create([Bind(Include = "Id,Title")] Section section)
+        public async Task<ActionResult> Create(
+            [Bind(Include = "Id,Title")] Section section)
         {
             if (ModelState.IsValid)
             {
                 db.Sections.Add(section);
-                db.SaveChanges();
+                await db.SaveChangesAsync();
                 return RedirectToAction("Index");
             }
 
@@ -60,13 +63,13 @@ namespace Memberships.Areas.Admin.Controllers
         }
 
         // GET: Admin/Section/Edit/5
-        public ActionResult Edit(int? id)
+        public async Task<ActionResult> Edit(int? id)
         {
             if (id == null)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
             }
-            Section section = db.Sections.Find(id);
+            Section section = await db.Sections.FindAsync(id);
             if (section == null)
             {
                 return HttpNotFound();
@@ -76,28 +79,29 @@ namespace Memberships.Areas.Admin.Controllers
 
         // POST: Admin/Section/Edit/5
         // To protect from overposting attacks, please enable the specific properties you want to bind to, for 
-        // more details see https://go.microsoft.com/fwlink/?LinkId=317598.
+        // more details see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public ActionResult Edit([Bind(Include = "Id,Title")] Section section)
+        public async Task<ActionResult> Edit(
+            [Bind(Include = "Id,Title")] Section section)
         {
             if (ModelState.IsValid)
             {
                 db.Entry(section).State = EntityState.Modified;
-                db.SaveChanges();
+                await db.SaveChangesAsync();
                 return RedirectToAction("Index");
             }
             return View(section);
         }
 
         // GET: Admin/Section/Delete/5
-        public ActionResult Delete(int? id)
+        public async Task<ActionResult> Delete(int? id)
         {
             if (id == null)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
             }
-            Section section = db.Sections.Find(id);
+            Section section = await db.Sections.FindAsync(id);
             if (section == null)
             {
                 return HttpNotFound();
@@ -108,11 +112,15 @@ namespace Memberships.Areas.Admin.Controllers
         // POST: Admin/Section/Delete/5
         [HttpPost, ActionName("Delete")]
         [ValidateAntiForgeryToken]
-        public ActionResult DeleteConfirmed(int id)
+        public async Task<ActionResult> DeleteConfirmed(int id)
         {
-            Section section = db.Sections.Find(id);
-            db.Sections.Remove(section);
-            db.SaveChanges();
+            Section section = await db.Sections.FindAsync(id);
+            var isUnused = await db.Items.CountAsync(i => i.SectionId.Equals(id)) == 0;
+            if (isUnused)
+            {
+                db.Sections.Remove(section);
+                await db.SaveChangesAsync();
+            }
             return RedirectToAction("Index");
         }
